1. feat: Signed In / Signed Out environments implemented 2. feat: Sign out added
1.
in store/index.js has been setted a getter ( isAuthenticated() ) which returns if current
user is signed in ( returns true ) or not ( returns false ).

In main.js has been implemented the Firebase method accountable to return user state in Firebase
and dispatch it to Vuex, who will handle the response.

App.vue will handle what options will be displayed according to if user is authenticated or not.

2.
In App.vue sign out button has been added, and according to if user is signed in or not, it`ll being displayed